@@ -2,7 +2,8 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <PlatformTarget>x64</PlatformTarget>
   </PropertyGroup>
 
   <ItemGroup>
@@ -42,28 +43,28 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="CsvHelper" Version="12.1.2" />
-    <PackageReference Include="Microsoft.AspNetCore.SignalR.Client" Version="1.1.0" />
-    <PackageReference Include="Microsoft.Extensions.Configuration" Version="3.1.4" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="3.1.4" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.FileExtensions" Version="3.1.4" />
+    <PackageReference Include="CsvHelper" Version="27.2.1" />
+    <PackageReference Include="Microsoft.AspNetCore.SignalR.Client" Version="6.0.1" />
+    <PackageReference Include="Microsoft.Extensions.Configuration" Version="6.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="6.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.FileExtensions" Version="6.0.0" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
-    <PackageReference Include="obs-websocket-dotnet" Version="4.2.0-beta3" />
+    <PackageReference Include="obs-websocket-dotnet" Version="4.9.1" />
     <PackageReference Include="Profanity.Detector" Version="0.1.8" />
-    <PackageReference Include="TwitchLib" Version="3.1.1" />
-    <PackageReference Include="TwitchLib.Api" Version="3.1.3" />
-    <PackageReference Include="TwitchLib.Api.Core" Version="3.1.3" />
-    <PackageReference Include="TwitchLib.Api.Core.Enums" Version="3.1.2" />
-    <PackageReference Include="TwitchLib.Api.Core.Interfaces" Version="3.1.2" />
-    <PackageReference Include="TwitchLib.Api.Core.Models" Version="3.1.3" />
-    <PackageReference Include="TwitchLib.Api.Helix" Version="3.1.3" />
-    <PackageReference Include="TwitchLib.Api.Helix.Models" Version="3.1.3" />
-    <PackageReference Include="TwitchLib.Api.V5" Version="3.1.2" />
-    <PackageReference Include="TwitchLib.Client" Version="3.1.5" />
-    <PackageReference Include="TwitchLib.Client.Enums" Version="3.1.4" />
-    <PackageReference Include="TwitchLib.Client.Models" Version="3.1.5" />
+    <PackageReference Include="TwitchLib" Version="3.3.0" />
+    <PackageReference Include="TwitchLib.Api" Version="3.4.1" />
+    <PackageReference Include="TwitchLib.Api.Core" Version="3.4.1" />
+    <PackageReference Include="TwitchLib.Api.Core.Enums" Version="3.4.1" />
+    <PackageReference Include="TwitchLib.Api.Core.Interfaces" Version="3.4.1" />
+    <PackageReference Include="TwitchLib.Api.Core.Models" Version="3.4.1" />
+    <PackageReference Include="TwitchLib.Api.Helix" Version="3.4.1" />
+    <PackageReference Include="TwitchLib.Api.Helix.Models" Version="3.4.1" />
+    <PackageReference Include="TwitchLib.Api.V5" Version="3.4.1" />
+    <PackageReference Include="TwitchLib.Client" Version="3.2.4" />
+    <PackageReference Include="TwitchLib.Client.Enums" Version="3.2.4" />
+    <PackageReference Include="TwitchLib.Client.Models" Version="3.2.4" />
     <PackageReference Include="TwitchLib.Communication" Version="1.0.3" />
-    <PackageReference Include="TwitchLib.PubSub" Version="3.1.4" />
+    <PackageReference Include="TwitchLib.PubSub" Version="3.2.3" />
   </ItemGroup>
 
   <ItemGroup>
@@ -87,9 +88,6 @@
   <ItemGroup>
     <Reference Include="ObsControl">
       <HintPath>..\ObsControl\bin\Debug\ObsControl.dll</HintPath>
-    </Reference>
-    <Reference Include="System">
-      <HintPath>System</HintPath>
     </Reference>
   </ItemGroup>
 
