@@ -11,6 +11,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using OverlayManager.Hubs;
 using Microsoft.Extensions.FileProviders;
+using Microsoft.Extensions.Hosting;
 
 namespace OverlayManager
 {
@@ -33,13 +34,14 @@ namespace OverlayManager
 				options.MinimumSameSitePolicy = SameSiteMode.None;
 			});
 
-			services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
+			services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_3_0);
 			services.AddSignalR(o => o.EnableDetailedErrors = true);
 			services.AddHostedService<BackgroundWorker>();
+			services.AddMvc(options => options.EnableEndpointRouting = false);
 		}
 
 		// This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-		public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+		public void Configure(IApplicationBuilder app, IHostEnvironment env)
 		{
 			if (env.IsDevelopment())
 			{
@@ -60,12 +62,14 @@ namespace OverlayManager
 				ServeUnknownFileTypes = true
 			});
 
+			app.UseRouting();
+
 			app.UseCookiePolicy();
 
-			app.UseSignalR(x => {
+			app.UseEndpoints(endpoints => {
 				// Browser connects here.
-				x.MapHub<MrAnnouncerBotHub>("/MrAnnouncerBotHub");
-				x.MapHub<CodeRushedHub>("/CodeRushedHub");
+				endpoints.MapHub<MrAnnouncerBotHub>("/MrAnnouncerBotHub");
+				endpoints.MapHub<CodeRushedHub>("/CodeRushedHub");
 			});
 			app.UseMvc();
 		}
