@@ -10,8 +10,7 @@ namespace DHDM
 	public class HealthDamageCommand : BaseStreamDeckCommand, IDungeonMasterCommand
 	{
 		int healthDamageValue;
-		bool changeAllPlayers;
-
+		
 		public void Execute(IDungeonMasterApp dungeonMasterApp, ChatMessage chatMessage)
 		{
 			DamageHealthChange damageHealthChange = new DamageHealthChange();
@@ -32,10 +31,8 @@ namespace DHDM
 		public bool Matches(string message)
 		{
 			Match match = Regex.Match(message, @"(^[\-\+]\d+)" + PlayerSpecifier);
-			changeAllPlayers = false;
 			if (!match.Success)
 			{
-				changeAllPlayers = true;
 				match = Regex.Match(message, @"(^[\-\+]\d+)");
 			}
 			if (match.Success)
